SLING-10236 : Introduce Operation.asRepoInitString (incorporate SLING-10192, thanks Karl for spotting)
