+++ src/main/java/org/apache/sling/repoinit/parser/operations/AclGroupBase.java
@@ -96,22 +96,6 @@ abstract class AclGroupBase extends Operation {
     }
 
     @NotNull
-    static String pathsToString(@NotNull List<String> paths) {
-        return listToString(paths.stream()
-                .map(s -> {
-                    if (s.startsWith(":") && s.contains("#")) {
-                        String func = s.substring(1, s.indexOf(":",1));
-                        String s2 = s.substring(func.length()+2, s.lastIndexOf('#'));
-                        String trailingPath = (s.endsWith("#")) ?  "" : s.substring(s.indexOf("#")+1);
-                        return func + "(" + s2 +")" + trailingPath;
-                    } else {
-                        return s;
-                    }
-                })
-                .collect(Collectors.toList()));
-    }
-
-    @NotNull
     private static String nodetypesToString(@NotNull List<String> nodetypes) {
         return (nodetypes.isEmpty()) ? "" : " nodetypes " + listToString(nodetypes);
     }
